New list format (updated 2)

+++ BtMatuoklis/app/src/main/java/com/example/btmatuoklis/DevInfo.java
@@ -3,17 +3,17 @@ package com.example.btmatuoklis;
 public class DevInfo {
     public String Name;
     public String Mac;
-    public short Rssi;
+    public byte Rssi;
 
     public DevInfo(){}
 
-    public DevInfo(String name, String mac, short rssi){
+    public DevInfo(String name, String mac, byte rssi){
         this.Name = name;
         this.Mac = mac;
         this.Rssi = rssi;
     }
 
-    public void updateRssi(short rssi){
+    public void updateRssi(byte rssi){
         this.Rssi = rssi;
     }
